refactor(tmdb client): move tmdb library setup and URLs to a it's own, separate file
